AmphiDEB.jl

Installation

AmphiDEB depends on EcotoxSystems.jl, which is curently not registered. Installation thus requires

using Pkg
Pkg.add(url="htts://github.com/simonhansul/EcotoxSystems.jl")
Pkg.add(url="htts://github.com/simonhansul/AmphiDEB")

Guided examples are given in the examples directory.

Notes on notation and terminology

Simplified DEB notation

For DEB-TKTD parameters and state variables, we use a slightly simplified notation, following these rules:

Leading capitalized character indicates a state variable

X: External resource

X_emb: Embryonic buffer

I: Ingested resource

A: Assimilates

M: Somatic maintenance costs

H: Maturity

J: Maturity maintenance costs

R: Reproduction buffer

k_X is a rate constant related to state variable X

eta_XY is the efficiency of transforming X into Y

X_int is the initial value of X

y_X indicates the response to X (e.g. chemical stressor or temperature). The relative response is always defined as a value that can be multiplied with the relevant term in the derivatives without further transformation.

Furthermore, surface areas and volumes are only given up to a proportionality:

surface area $\propto$ mass $^{2/3}$

volume $\propto$ mass

The zoom factor is based on the maximum structural masses of individuals.

Life stage specificity

Some parameters are life stage specific, indicated by suffixes _emb, _lrv, _mt, _juv, _ad. <br> The suffix indicates the life stage for which a switch in the value occurrs. <br> For example, we have values k_M_emb and k_M_juv, which means that k_M has the value k_M_emb for embryos, larvae and metamorphs, and value k_M_juv for juveniles and adults.

Input parameters

Listed below are the input parameters of the model (valid for v0.3.3). <br> For guidance on how to use these, please consult the examples. <br> Please note that the default parameters are largely arbitrary and do not reflect a particular species.

Global (glb)

Parameter	Default value	Comment

t_max	56.	Maximum simulation time

N0	1.	Number of starting individuals

dX_in	[20., 20.]	Food input rate [mg d⁻¹], assuming ad libitum feeding

k_V	[0., 0.]	Dilution rate in the aquatic medium[d⁻¹]

V_patch	[1., 1.]	Simulated volume and surface area of the aquatic and terrestic habitat, respectively

T	293.15	Ambient temperature [K]

C_W	[0.;]	Exposure concentrations (matrix: treatments × compounds)

pathogeninoculationdose	0.	Pathogen spores added to medium [# spores]

pathogeninoculationtime	30.	Time-point of pathogen inoculation [t]

medium_renewals	[0.]	Time-points for media renewals (removes spores)

Species-specific (spc)

Parameter	Default value	Comment

Metaparameters

Z	Dirac(1.)	zoom factor, based on maximum structural masses

propagate.zoom.dImaxemb	1/3	dI ∝ Z^(1/3)

propagate.zoom.dImaxlrv	1/3

propagate.zoom.dImaxjuv	1/3

propagate.zoom.Xembint	1.	X_emb ∝ Z

propagate.zoom.H_j1	1.	H ∝ Z

propagate.zoom.H_p	1.

propagate.zoom.KXlrv	1/3	K_X ∝ dI ∝ Z^(1/3)

propagate.zoom.KXjuv	1/3

Physiological baseline (DEB)

Xembint	1	Initial vitellus (≈ dry mass of an egg)

KXlrv	1.	Larval half-saturation constant for food uptake

KXjuv	1.	Juvenile/adult half-saturation constant for food uptake

dImaxemb	1	Embryonic maximum specific ingestion rate

dImaxlrv	1	Larval maximum specific ingestion rate

dImaxjuv	1	Juvenile/adult maximum specific ingestion rate

kappa_emb	0.8	Embryonic to metamorph allocation fraction

kappa_juv	0.8	Juvenile/adult allocation fraction to soma

gamma	0.5	Larval allocation to metamorphic reserves

eta_IA	0.54	Assimilation efficiency

etaASemb	0.4	Embryonic growth efficiency

etaASjuv	0.4	Juvenile/adult growth efficiency

eta_AR	0.95	Reproduction efficiency

eta_SA	0.8	Shrinking efficiency

kMemb	0.11	Embryonic somatic maintenance rate

kMjuv	0.11	Juvenile/adult somatic maintenance rate

kJemb	0.027	Embryonic maturity maintenance rate

kJjuv	0.027	Juvenile/adult maturity maintenance rate

H_j1	1	Maturity at start of metamorphosis

H_p	55.	Maturity at puberty

T_A	8000.	Arrhenius temperature (K)

T_ref	293.15	Reference temperature

b_T	40.	Temperature effect on resource allocation (cf. Romoli et al. (2025))

TKTD Parameters

h_b	0.	Background mortality

KD	[0. 0. 0. 0. 0. 0. 0.;]	k_D per PMoA and stressor

B	[2. 2. 2. 2. 2. 2. 2.;]	Slope parameters

E	[1e10 1e10 1e10 1e10 1e10 1e10 1e10;]	Median effective damage per stressor and PMoA

KD_h	[0.;]	k_D for GUTS-SD module per stressor

E_h	[1e10;]	Sensitivity for GUTS-SD

B_h	[1.;]	Slope for GUTS-SD

C_h	[1.;]	Proportionqlity constant to convert larval sublethal response to post-metamorphic hazard rate

Srelcrit	0.66	Critical relative body mass loss

h_S	0.6	Hazard rate below critical mass

a_max	truncated(Normal(15*365, 1.5*365), 0, Inf)	Maximum age [d]

tau_R	365.	Reproduction period [d]

Pathogen effects

Chi	LogNormal(log(1)+1^2, 1)	Killing rate modifier (log-normal with mode 1, σ=1)

E_P	[Inf, Inf, Inf, Inf]	Sensitivity thresholds for pathogen effects

B_P	[2., 2., 2., 2.]	Slope parameters for pathogen effects

Pathogens (pth)

The model is adopted from Drawert et al. (2017). Default values are the averages of the suggested ranges.

Parameter	Default value	Comment

gamma	geomean([1e-6, 1.])	Zoospore encounter rate [spores individual⁻¹ d⁻¹]

eta	harmmean([5., 20.])	Zoospore production rate

v0	0.5	Zoospore encystment rate

f	0.5	Host reinfection fraction

sigma0	harmmean([0.1, 0.5])	Sporangia killing rate

sigma1	harmmean([0.1, 0.5])	Density-dependent killing rate

mu	harmmean([0.01, 1.5])	Zoospore death rate

API

AmphiDEB.AmphiDEB_ODE_with_linear_TD! — Method

AmphiDEB_ODE_with_linear_TD!(du, u, p, t)::Nothing

ODE system with linear toxicodynamics.

AmphiDEB.AmphiDEB_ODE_with_loglogistic_TD! — Method

AmphiDEB_ODE_with_loglogistic_TD!(du, u, p, t)::Nothing

ODE system with log-logistic toxicodynamics.

AmphiDEB.AmphiDEB_global_rules! — Method

AmphiDEB_global_rules!(m)

Global rule-based portion of the AmphiDEB IBM.

AmphiDEB.Amphibian_DEB! — Method

Amphibian_DEB!(du, u, p, t)::Nothing

The default amphibian DEB model assumes a metamorphic reserve compartment, which is accumulated during larval development, and depleted during metamorphosis. During metamorphosis, ingestion rate decreases gradually and reaches 0 at the end of metamorphosis.

AmphiDEB.IBM_simulator — Method

IBM_simulator(
    p::ComponentVector;

    init_global_statevars = initialize_global_statevars,
    global_ode! = AmphiDEB_global!,
    global_rules! = AmphiDEB_global_rules!,
    
    init_individual_statevars = initialize_individual_statevars,
    individual_ode! = AmphiDEB_individual_ODE_with_loglogistic_TD!, 
    individual_rules! = default_individual_rules!,
    gen_ind_params = p -> EcotoxSystems.generate_individual_params(p; pth = p.pth),

    kwargs...
    )

Compose and execute an individual-based simulation from the given components. <br> This function is a simple wrapper around EcotoxSystems.IBM_simulator. <br> Its main purpose is to provide a default configuration for the AmphiDEB model.

Positional arguments

p: Parameters for all components: global (p.glb), Amphibian species-specific (p.spc) and pathogen-specific (p.pth)

Keyword arguments

Specification of global part of the system

init_global_statevars: Function that initializes global state variables as ComponentVector, e.g. food abundance, temperature.

global_ode!: Defintion of global ODE-based portion of the system (e.g. growth or experimental addition of food resource).

global_rules!: Definition of global rule-based portion of the system (e.g. computing population aggregates).

Specification of species/individual-specific part of the system

init_individual_statevars: Function that initializes individual-level state variables as ComponentVector, e.g. structural mass, scaled adamge, etc.

individual_ode!: Definition of individual-level ODE-based portion of the system (e.g. derivatives of the DEB-TKTD module).

individual_rules!: Definition of individual-level rule-based portion of the system (e.g. death, discrete reproduction events).

gen_ind_params: A function that converts species-specific parameters to individual-specific parameters. For parameters which are provided as distributions, this takes a random sample from the distribution each time an individual is generated.

This argument probably does not have to be changed, unless additional components (other than glb, spc and pth) are added to the system.

AmphiDEB.ODE_simulator — Method

ODE_simulator(
    p::ComponentVector; 
    model = AmphiDEB_ODE_M1!, 
    callbacks = AmphODE_callbacks(), 
    statevars_init = initialize_statevars,
    kwargs...
    )

A wrapper around the EcotoxSystems.jl ODE_simulator for application with the AmphiDEB model.

AmphiDEB.TKTD_LL2! — Method

TKTD_LL2!(du, u, p, t)::Nothing

TKTD model with following configuration:

Mixture toxicity based on independent action (IA)

Log-logistic relationship between damage and metabolic processes

AmphiDEB.TKTD_linear! — Method

TKTD_linear!(du, u, p, t)::Nothing

TKTD model with following configuration:

Mixture toxicity based on independent action (IA)

Linear-above-threshold relationship between damage and metabolic processes (the default DEBtox stress function)

AmphiDEB.calc_H_eq_juv — Method

calc_H_eq_juv(spc::ComponentVector)

Calculate equilibrium maturity for juvenile/adult stages.

AmphiDEB.calc_H_eq_lrv — Method

calc_H_eq_lrv(spc::ComponentVector)

Calculate equilibrium maturity for larval stage.

AmphiDEB.calc_S_B — Method

calc_S_B(sim::AbstractDataFrame)

Extracts structural mass at birth from simulation output.

AmphiDEB.calc_S_max_ad — Method

calc_S_max_ad(spc::ComponentVector)

Alias for calc_S_max_juv.

AmphiDEB.calc_S_max_juv — Method

calc_S_max_juv(spc::ComponentVector)

Calculates maximum structural mass of juveniles and adults from parameters.

AmphiDEB.calc_a_B — Method

calc_a_B(sim::AbstractDataFrame)

Extracts age at birth from simulation output.

AmphiDEB.calc_a_j1 — Method

calc_a_j1(sim::AbstractDataFrame)

Extract age at beginning of metamorphosis, j1, from simulation output. <br> Beginning of metamorphosis is defined as the point when feeding rates start to decline. <br> This can align with Gosner stage 42 for amphibians, but can also be as early as Gosner 38.

AmphiDEB.calc_a_j2 — Method

calc_a_j2(sim::AbstractDataFrame)

Extract age at the end of metamorphosis, based on depletion of metamorphic reserve. <br> Without providing an additional maturity threshold (and therefore an additional argument to this function), this is the only plausible option we currently have.<br> In that case, the end of metamorphosis is reached when the reserve is depleted.

AmphiDEB.calc_a_j2 — Method

calc_a_j2(sim::AbstractDataFrame, H_j2)

Extract age at the end of metamorphosis, based on a maturity threshold H_j2. <br>

AmphiDEB.calc_dE_mt_lrv — Method

calc_dE_mt_lrv(
    eta_AS::Float64, 
    y_G::Float64,
    y_G_P::Float64,
    gamma::Float64,
    kappa::Float64,
    dA::Float64,
    dM::Float64
    )::Float64

Accumulation of metamorphic reserve by larvae.

AmphiDEB.calc_dE_mt_mt — Method

calc_dE_mt_mt(
    dH::Float64,
    dJ::Float64,
    dM::Float64
    )::Float64

Depletion of metamorphic reserve by metamorphs.

AmphiDEB.calc_dH — Method

calc_dH(
    kappa::Float64,
    dA::Float64,
    dJ::Float64
    )::Float64

Maturation rate following standard DEBkiss model.

AmphiDEB.calc_dH_mt — Method

calc_dH_mt(
    kappa::Float64,
    dM::Float64,
    dJ::Float64
    )::Float64

Maturation rate for metamorphs, calculated from the κ-rule and assuming that soma are decoupled from the κ rule during metamorphosis.

AmphiDEB.calc_dI_emb — Method

calc_dI_emb(
    embryo::Float64,
    S::Float64,
    dI_max_emb::Float64,
    y_T::Float64
    )::Float64

Embryonic buffer uptake rate for embryos.

AmphiDEB.calc_dI_juv — Method

calc_dI_juv(
    juvenile::Float64,
    adult::Float64,
    f_X::Float64,
    dI_max_juv::Float64,
    S::Float64,
    y_T::Float64
    )::Float64

Food ingestion rate for juveniles and adults.

AmphiDEB.calc_dI_lrv — Method

calc_dI_lrv(
    larva::Float64,
    f_X::Float64,
    dI_max_lrv::Float64,
    S::Float64,
    y_T::Float64,
    )::Float64

Food ingestion rate for larvae.

AmphiDEB.calc_dI_mt — Method

calc_dI_mt(
    metamorph::Float64,
    f_X::Float64,
    dI_max_lrv::Float64,
    E_mt::Float64,
    E_mt_max::Float64,
    S::Float64,
    y_T::Float64
    )::Float64

Residual ingestion rate for metamorphs (if any).

AmphiDEB.calc_dS_emb_juv_ad — Method

calc_dS_emb_juv_ad(
    kappa::Float64,
    dA::Float64,
    dM::Float64,
    eta_SA::Float64,
    y_G::Float64,
    y_GP::Float64,
    eta_AS::Float64
    )::Float64

Somatic growth rate for embryos, juveniles and adults, including response to chemicals (y_G) and pathogens (y_GP).

AmphiDEB.calc_dS_lrv — Method

calc_dS_lrv(
    kappa::Float64,
    dA::Float64,
    dM::Float64,
    eta_SA::Float64,
    gamma::Float64,
    eta_AS::Float64,
    y_G::Float64,
    y_GP::Float64
    )::Float64

Somatic growth rate for larvae. Approach follows DEBkiss model, but with an additional γ-split to divert assimilates towards the metamorphic reserve. This is also taken into account in the shrinking equation (kappa * dA < dM).

AmphiDEB.calc_dS_mt — Method

calc_dS_mt(
    metamorph::Float64,
    eta_AS::Float64,
    y_G::Float64,
    y_GP::Float64,
    dA::Float64
    )::Float64

Somatic growth rate for metamorphs.

AmphiDEB.calc_eta_AS — Method

calc_eta_AS(
    embryo::Float64,
    larva::Float64,
    metamorph::Float64,
    eta_AS_emb::Float64,
    juvenile::Float64,
    adult::Float64,
    eta_AS_juv::Float64
    )::Float64

Individual's current value of the growth efficiency eta_AS, based on life stage indicators and life stage-specific values of eta_AS.

AmphiDEB.calc_kappa — Method

calc_kappa(
    embryo::Float64,
    larva::Float64,
    metamorph::Float64,
    kappa_emb::Float64,
    juvenile::Float64,
    adult::Float64,
    kappa_juv::Float64,
    b_T::Float64, 
    T_ref::Float64,
    y_K::Float64
    )::Float64

Individual's current value of the growth efficiency kappa, based on life stage indicators and life stage-specific values of kappa.

This includes temperature effects on κ as in Romoli et al. (2024).

AmphiDEB.calc_metamorphosis_duration — Method

calc_metamorphosis_duration(sim::AbstractDataFrame)

Calculate duration of metamorphosis.

AmphiDEB.calc_vB_growthrate_lrv — Method

calc_vB_growthrate_lrv(spc::ComponentVector)::Float64

Calculate larval von Bertalanffy growth rate.

AmphiDEB.dA — Method

dA(
    dI::Float64, 
    eta_IA::Float64, 
    y_A::Float64, 
    y_AP::Float64
    )::Float64

Assimilation flux, following standard DEBkiss model.

AmphiDEB.dE_mt — Method

dE_mt(
    larva::Float64,
    dE_mt_lrv::Float64,
    metamorph::Float64,
    dE_mt_mt::Float64
    )::Float64

Life stage-specific dynamics of the metamorphic reserve.

AmphiDEB.dE_mt_max — Method

dE_mt_max(
    larva::Float64,
    dE_mt::Float64
    )::Float64

Function to record the maximum metamorphic reserve level, reached at the transition from larva to metamorph.

AmphiDEB.dH — Method

dH(
    adult::Float64,
    metamorph::Float64,
    dH::Float64,
    dH_mt::Float64
    )::Float64

Life stage-specific maturation rate.

AmphiDEB.dI — Method

dI(
    dI_emb::Float64,
    dI_mt::Float64, 
    dI_lrv::Float64, 
    dI_juv_ad::Float64
    )::Float64

Food ingestion rate with account for life stage-specificity.

AmphiDEB.dJ — Method

dM(
    S::Float64,
    k_M::Float64,
    y_M::Float64,
    y_MP::Float64,
    y_T::Float64
    )::Float64

Maturity maintenance rates, including response to chemicals (y_M), pathogens (y_MP) and temperature (y_T).

AmphiDEB.dM — Method

dM(
    S::Float64,
    k_M::Float64,
    y_M::Float64,
    y_MP::Float64,
    y_T::Float64
    )::Float64

Somatic maintenance rate, including response to chemicals (y_M), pathogens (y_MP) and temperature (y_T).

AmphiDEB.dP_S — Method

function dP_S(
    v0::Float64,
    gamma::Float64,
    P_Z::Float64,
    eta::Float64,
    f::Float64,
    P_S::Float64,
    sigma0::Float64,
    sigma1::Float64,
    Chi::Float64
    )::Float64

Derivative of the individual-specific sporangia abundace P_S.

AmphiDEB.dP_Z — Method

function dP_Z(
    mu::Float64,
    P_Z::Float64
    )::Float64

Derivative of the zoospore abundance P_Z.

This function does not take changes in zoopsore abundance due to infection into account. This is handled in Pathogen_Infection!.

args:

mu: Zoospore background mortality rate

P_Z`: Current zoospore abundance in the environment

AmphiDEB.dR — Method

dR(
    adult::Float64,
    eta_AR::Float64,
    y_R::Float64,
    y_RP::Float64,
    kappa::Float64,
    dA::Float64,
    dJ::Float64
    )::Float64

Reproduction rate, including response to chemicals y_R and pathogens (y_RP).

AmphiDEB.dS — Method

dS(
    embryo::Float64,
    juvenile::Float64,
    adult::Float64,
    dS_emb_juv_ad::Float64,
    larva::Float64,
    dS_lrv::Float64,
    metamorph::Float64,
    dS_mt::Float64
    )::Float64

Life stage-specific somatic growth rate.

AmphiDEB.determine_k_J — Method

determine_k_J(
    embryo::Float64,
    larva::Float64,
    metamorph::Float64,
    k_J_emb::Float64,
    juvenile::Float64,
    adult::Float64,
    k_J_juv::Float64
    )::Float64

Current value of k_J, depending on life stage.

AmphiDEB.determine_k_M — Method

determine_k_M(
    embryo::Float64,
    larva::Float64,
    metamorph::Float64,
    k_M_emb::Float64,
    juvenile::Float64,
    adult::Float64,
    k_M_juv::Float64
    )::Float64

Current value of k_M, depending on life stage.

AmphiDEB.determine_life_stage! — Method

determine_life_stage!(du, u, p, t)::Nothing

Determine an individual's current life stages by calculating all life stage indicators. To avoid numerical issues during ODE solve, all life stage transitions are approximated by sigmoid functions.

AmphiDEB.f_X — Method

f_X(
    larva::Float64,
    metamorph::Float64, 
    juvenile::Float64, 
    adult::Float64,
    X::Vector{Float64},
    V_patch::Union{Vector{Real},Vector{Float64}}, 
    K_X_lrv::Float64,
    K_X_juv::Float64
    )::Float64

Scaled functional response with account for life stage-specific half saturation cosntant K_X.

AmphiDEB.f_X — Method

f_X(
    X::Float64,
    V_patch::Float64,
    K_X::Float64
    )::Float64

Scaled functional response f_X based on Holling Type II functional response.

AmphiDEB.initialize_individual_statevars — Method

" initializeindividualstatevars(p::ComponentVector; kwargs...)::ComponentVector

Initialize individual-level state variables for the AmhpiDEB model. Additional states can be added via kwargs. If existing states are provided via kwargs, these will be overwritten.

IMPORTANT NOTE FOR SIMULATING MIXTURES: You can simulate an arbitrary number of chemical stressors, but it is currently not possible to dynamically change the shape of vectors and matrices contained in a ComponentVector. In practice, this means: If you want to simulate mixtures, you cannot simply provide the parameters as kwargs to this function, but the entire ComponentVector has to be re-defined. You can do so by copy-pasting the definition body of this function and changing the shape of y_j. The same is true for the parameter vector, where the shape of all TKTD parameters has to be adjusted. An example is given in the unit tests of the EcotoxSystems package: https://github.com/SimonHansul/EcotoxSystems.jl/blob/main/test/test05_mixtures.jl For application to the AmphiDEB model, we have to take into account that it has an additional PMoA, and therefore an additional column in the sublethal TKTD parameters.

AmphiDEB.is_adult — Method

is_adult(
    H::Float64,
    H_p::Float64;
    beta = 1e3
    )::Float64

Determine wheter an individual is in adult state, returning life stage indicator as float. Adults are individuals with a maturity level above the puberty threshold H_p.

AmphiDEB.is_embryo — Method

is_embryo(
    X_emb::Float64;
    beta::Float64 = 1e3
    )::Float64

Determine whether an individual is in embryonic state, returning life stage indicator as float. Following rules of the standard DEBkiss model, embryonic state is maintained until the embryonic buffer is emptied (X_emb_int=0).

AmphiDEB.is_juvenile — Method

is_juvenile(
    H::Float64,
    H_j1::Float64,
    y_H_neg::Float64,
    y_H_pos::Float64,
    E_mt::Float64,
    H_p::Float64;
    beta1 = 1e3,
    beta2 = 1e3,
    beta3 = 1e3
    )::Float64

Determine wheter an individual is in juvenile state, returning life stage indicator as float. Juveniles are individuals with a maturity level between H_j1 and H_p whose metamorphic reserve E_mthas been emptied.

AmphiDEB.is_larva — Method

is_larva(
    X_emb::Float64,
    H::Float64,
    H_j1::Float64,
    y_H_neg::Float64, 
    y_H_pos::Float64;
    beta1::Float64 = 1e3,
    beta2::Float64 = 1e7
    )::Float64

Determine wheter an individual is in larval state, returning life stage indicator as float. Larvae are all non-embryos whose maturity is below the threshold for metamorphosis H_j1. The timing of H_j1 will often be aligned with Gosner stage 42 for anurans, but might ocurr earlier. The decisive factor is the decline in feeding rates.

AmphiDEB.is_metamorph — Method

is_metamorph(
    H::Float64, 
    H_j1::Float64, 
    y_H_neg::Float64,
    y_H_pos::Float64,
    E_mt::Float64;
    beta1::Float64 = 1e7, 
    beta2::Float64 = 1e3, 
    )::Float64

Determine wheter an individual is in metamorph state, returning life stage indicator as float. Metamorphs have a maturity level above `H_j1, and their metamorphic reserve E_mt is not emptied yet.

AmphiDEB.life_stage_and_plasticity_effects — Method

life_stage_and_plasticity_effects(du, u, p, t)::Tuple{Float64,Float64}

Life-stage specificity of parameters., as well as plastic responses to environmental factors (e.g. effect of temperature on κ).

AmphiDEB.minimal_TK_aquatic — Method

minimal_TK_aquatic(
    larva::Float64,
    k_D::Float64, 
    C_W::Float64,
    D::Float64
    )::Float64

Minimal TK model (no feedbacks) for aquatic exposure. Only applies to larvae.

AmphiDEB.temperature_sinusoidal — Method

temperature_sinusoidal(t::Float64, T_max::Float64, T_min::Float64, t_peak ::Float64)::Float64

Calculate seasonal fluctuations in temperature from a sinusoidal function.

AmphiDEB.y_T — Method

y_T(
    T_A::Float64,
    T_ref::Float64,
    T::Float64
    )::Float64

Temperature correction coefficient y_T according to Arrhenius equation.