AmphiDEB.jl

Installation

AmphiDEB depends on EcotoxSystems.jl, which is curently not registered. Installation thus requires

using Pkg
Pkg.add(url="htts://github.com/simonhansul/EcotoxSystems.jl")
Pkg.add(url="htts://github.com/simonhansul/AmphiDEB")

TKTD interface

The TKTD module is designed so that mixtures with an arbitrary number of chemicals can be simulated, and each chemical can act via arbitrary combinations of PMoAs. <br> For this pupose, TKTD parameters for sublethal effects are stored in matrices, where the rows represent chemicals and the columns represent PMoAs. <br> This affects the parameters $k_{D_j}$, $e_j$ and $b_j$. <br> The PMoAs have a fixed order in the parameter matrices:

Decrease in growth efficiency ($G$)

Increase in somatic and maturity maintenance costs ($M$)

Decrease in assimilation efficiency ($A$)

Decrease in reproduction efficiency ($R$)

Decrease in maturity threshold for metamorphosis ($H^-$)

Increase in maturity threshold for metamorphosis ($H^+$)

Decrease in $\kappa$ ($\kappa$, acceleration of ontogenesis) ($\Kappa$)

API

AmphiDEB.AmphiDEB_ODE_alt! — Method

Amphibian_DEB_alt!(du, u, p, t)::Nothing

Complete ODE system for alternative formulation of the AmphiDEB model (see Amphibian_DEB_alt!)

AmphiDEB.AmphiDEB_ODE_with_linear_TD! — Method

AmphiDEB_ODE_with_linear_TD!(du, u, p, t)::Nothing

ODE system with liner toxicodynamics.

AmphiDEB.AmphiDEB_ODE_with_loglogistic_TD! — Method

AmphiDEB_ODE_with_loglogistic_TD!(du, u, p, t)::Nothing

ODE system with log-logistic toxicodynamics.

AmphiDEB.AmphiDEB_global_rules! — Method

AmphiDEB_global_rules!(m)

Global rule-based portion of the AmphiDEB IBM.

AmphiDEB.Amphibian_DEB! — Method

Amphibian_DEB!(du, u, p, t)::Nothing

The default amphibian DEB model assumes a metamorphic reserve compartment, which is accumulated during larval development, and depleted during metamorphosis. During metamorphosis, ingestion rate decreases gradually and reaches 0 at the end of metamorphosis.

AmphiDEB.Amphibian_DEB_alt! — Method

Amphibian_DEB_alt!(du, u, p, t)::Nothing

Alternative formulation of the AmphiDEB model, where E_mt is viewed as a sub-compartment of structure. This leads to the following main differences:

E_mt is subject to somatic maintenance

E_mt contributes to the surface area scaling factor

E_mt is affected by growth efficiency, as well as effects on growth efficiency

AmphiDEB.IBM_simulator — Method

IBM_simulator(
    p::ComponentVector;

    init_global_statevars = initialize_global_statevars,
    global_ode! = AmphiDEB_global!,
    global_rules! = AmphiDEB_global_rules!,
    
    init_individual_statevars = initialize_individual_statevars,
    individual_ode! = AmphiDEB_individual_ODE_with_loglogistic_TD!, 
    individual_rules! = default_individual_rules!,
    gen_ind_params = p -> EcotoxSystems.generate_individual_params(p; pth = p.pth),

    kwargs...
    )

Compose and execute an individual-based simulation from the given components. <br> This function is a simple wrapper around EcotoxSystems.IBM_simulator. <br> Its main purpose is to provide a default configuration for the AmphiDEB model.

Positional arguments

p: Parameters for all components: global (p.glb), Amphibian species-specific (p.spc) and pathogen-specific (p.pth)

Keyword arguments

Specification of global part of the system

init_global_statevars: Function that initializes global state variables as ComponentVector, e.g. food abundance, temperature.

global_ode!: Defintion of global ODE-based portion of the system (e.g. growth or experimental addition of food resource).

global_rules!: Definition of global rule-based portion of the system (e.g. computing population aggregates).

Specification of species/individual-specific part of the system

init_individual_statevars: Function that initializes individual-level state variables as ComponentVector, e.g. structural mass, scaled adamge, etc.

individual_ode!: Definition of individual-level ODE-based portion of the system (e.g. derivatives of the DEB-TKTD module).

individual_rules!: Definition of individual-level rule-based portion of the system (e.g. death, discrete reproduction events).

gen_ind_params: A function that converts species-specific parameters to individual-specific parameters. For parameters which are provided as distributions, this takes a random sample from the distribution each time an individual is generated.

This argument probably does not have to be changed, unless additional components (other than glb, spc and pth) are added to the system.

AmphiDEB.ODE_simulator — Method

ODE_simulator(
    p::ComponentVector; 
    model = AmphiDEB_ODE_M1!, 
    callbacks = AmphODE_callbacks(), 
    statevars_init = initialize_statevars,
    kwargs...
    )

A wrapper around the EcotoxSystems.jl ODE_simulator for application with the AmphiDEB model.

AmphiDEB.TKTD_LL2! — Method

TKTD_LL2!(du, u, p, t)::Nothing

TKTD model with following configuration:

Mixture toxicity based on independent action (IA)

Log-logistic relationship between damage and metabolic processes

AmphiDEB.TKTD_linear! — Method

TKTD_linear!(du, u, p, t)::Nothing

TKTD model with following configuration:

Mixture toxicity based on independent action (IA)

Linear-above-threshold relationship between damage and metabolic processes (the default DEBtox stress function)

AmphiDEB.calc_H_eq_juv — Method

calc_H_eq_juv(spc::ComponentVector)

Calculate equilibrium maturity for juvenile/adult stages.

AmphiDEB.calc_H_eq_lrv — Method

calc_H_eq_lrv(spc::ComponentVector)

Calculate equilibrium maturity for larval stage.

AmphiDEB.calc_S_B — Method

calc_S_B(sim::AbstractDataFrame)

Extracts structural mass at birth from simulation output.

AmphiDEB.calc_S_max_ad — Method

calc_S_max_ad(spc::ComponentVector)

Alias for calc_S_max_juv.

AmphiDEB.calc_S_max_juv — Method

calc_S_max_juv(spc::ComponentVector)

Calculates maximum structural mass of juveniles and adults from parameters.

AmphiDEB.calc_a_B — Method

calc_a_B(sim::AbstractDataFrame)

Extracts age at birth from simulation output.

AmphiDEB.calc_a_j1 — Method

calc_a_j1(sim::AbstractDataFrame)

Extract age at beginning of metamorphosis, j1, from simulation output. <br> Beginning of metamorphosis is defined as the point when feeding rates start to decline. <br> This can align with Gosner stage 42 for amphibians, but can also be as early as Gosner 38.

AmphiDEB.calc_a_j2 — Method

calc_a_j2(sim::AbstractDataFrame)

Extract age at the end of metamorphosis, based on depletion of metamorphic reserve. <br> Without providing an additional maturity threshold (and therefore an additional argument to this function), this is the only plausible option we currently have.<br> In that case, the end of metamorphosis is reached when the reserve is depleted.

AmphiDEB.calc_a_j2 — Method

calc_a_j2(sim::AbstractDataFrame, H_j2)

Extract age at the end of metamorphosis, based on a maturity threshold H_j2. <br>

AmphiDEB.calc_kappa — Method

calc_kappa(
    embryo::Float64,
    larva::Float64,
    metamorph::Float64,
    kappa_emb::Float64,
    juvenile::Float64,
    adult::Float64,
    kappa_juv::Float64,
    b_T::Float64, 
    T_ref::Float64,
    y_K::Float64
    )::Float64

Calculate kappa accounting for life-stage, temperature effects and chemical effects.

Temperature effects are implemented according to Romoli et al. (2024).

AmphiDEB.calc_metamorphosis_duration — Method

calc_metamorphosis_duration(sim::AbstractDataFrame)

Calculate duration of metamorphosis.

AmphiDEB.calc_vB_growthrate_lrv — Method

calc_vB_growthrate_lrv(spc::ComponentVector)::Float64

Calculate larval von Bertalanffy growth rate.

AmphiDEB.dP_S — Method

function dP_S(
    v0::Float64,
    gamma::Float64,
    P_Z::Float64,
    eta::Float64,
    f::Float64,
    P_S::Float64,
    sigma0::Float64,
    sigma1::Float64,
    Chi::Float64
    )::Float64

Derivative of the individual-specific sporangia abundace P_S.

AmphiDEB.dP_Z — Method

function dP_Z(
    mu::Float64,
    P_Z::Float64
    )::Float64

Derivative of the zoospore abundance P_Z.

This function does not take changes in zoopsore abundance due to infection into account. This is handled in Pathogen_Infection!.

args:

mu: Zoospore background mortality rate

P_Z`: Current zoospore abundance in the environment

AmphiDEB.growth! — Method

growth!(du, u, p, t)

Calculation of life stage-specific growth fluxes for amphibians. <br>

AmphiDEB.growth_alt! — Method

growth_alt!(du, u, p, t)::Tuple{Real,Real}

Calculation of growth fluxes for amphibians. <br> Returns life stage-specific values for eta_AS and kappa.

AmphiDEB.ingestion! — Method

ingestion!(du, u, p, t)::Nothing

Life stage-specific calculation of ingestion rate for amphibians. <br> Note that this function requires all components to be included in the arguments du, u and p, so that we can access the external food concentration. <br> Hence, we have u.ind, u.glb, p.ind, etc., instead of simply u and p.

AmphiDEB.ingestion_alt! — Method

ingestion!(du, u, p, t)::Nothing

Life stage-specific calculation of ingestion rate for amphibians. <br> Note that this function requires all components to be included in the arguments du, u and p, so that we can access the external food concentration. <br> Hence, we have u.ind, u.glb, p.ind, etc., instead of simply u and p.

AmphiDEB.initialize_individual_statevars — Method

" initializeindividualstatevars(p::ComponentVector; kwargs...)::ComponentVector

Initialize individual-level state variables for the AmhpiDEB model. Additional states can be added via kwargs. If existing states are provided via kwargs, these will be overwritten.

IMPORTANT NOTE FOR SIMULATING MIXTURES: You can simulate an arbitrary number of chemical stressors, but it is currently not possible to dynamically change the shape of vectors and matrices contained in a ComponentVector. In practice, this means: If you want to simulate mixtures, you cannot simply provide the parameters as kwargs to this function, but the entire ComponentVector has to be re-defined. You can do so by copy-pasting the definition body of this function and changing the shape of y_j. The same is true for the parameter vector, where the shape of all TKTD parameters has to be adjusted. An example is given in the unit tests of the EcotoxSystems package: https://github.com/SimonHansul/EcotoxSystems.jl/blob/main/test/test05_mixtures.jl For application to the AmphiDEB model, we have to take into account that it has an additional PMoA, and therefore an additional column in the sublethal TKTD parameters.

AmphiDEB.life_stage_and_plasticity_effects — Method

life_stage_and_plasticity_effects(du, u, p, t)::Tuple{Float64,Float64}

Handles life-stage specificity of parameters. The parameter notation foresees that the superscript indicates the first life stage for which a value is valid, e.g. if we have etaASemb and etaASjuv, then etaASemb is applied for eymbros, larvae and metamorphs, and etaASjuv is applied for juveniles and adults

AmphiDEB.maintenance! — Method

maintenance!(du, u, p, t)::Nothing

Calculation of maintenance fluxes for amphibians.

AmphiDEB.maintenance_alt! — Method

maintenance!(du, u, p, t)::Nothing

Calculation of maintenance fluxes for amphibians.

AmphiDEB.maturation! — Method

maturation!(du, u, p, t, kappa)::Nothing

Calculation of maturation fluxes for amphibians.

AmphiDEB.metamorphic_reserve! — Method

metamorphic_reserve!(du, u, p, t, kappa)::Nothing

Calculation of metamorphic reserve dynamics for amphibians. <br> Metamorphic reserve is accumulated during larval development and depleted during metamorphic climax.

AmphiDEB.metamorphic_reserve_alt! — Method

metamorphic_reserve!(du, u, p, t, kappa)::Nothing

Calculation of metamorphic reserve dynamics for amphibians. <br> Metamorphic reserve is accumulated during larval development and depleted during metamorphic climax.

AmphiDEB.minimal_TK_aquatic — Method

minimal_TK_aquatic(
    larva::Float64,
    k_D::Float64, 
    C_W::Float64,
    D::Float64
    )::Float64

Minimal TK model (no feedbacks) for aquatic exposure. Only applies to larvae.

AmphiDEB.temperature_sinusoidal — Method

temperature_sinusoidal(t::Float64, T_max::Float64, T_min::Float64, t_peak ::Float64)::Float64

Calculate seasonal fluctuations in temperature from a sinusoidal function.

AmphiDEB.y_T — Method

y_T(
    T_A::Float64,
    T_ref::Float64,
    T::Float64
    )::Float64

Calculates temperature correction coefficient y_T according to Arrhenius equation.

T_A : Arrhenius temperature (K)

T_ref : Reference temperature (K)

T : Current ambient temperature